+++ src/Dropdown/Dropdown.test.js
@@ -1,10 +1,12 @@
 import { Button } from '../';
 import { Dropdown } from './Dropdown';
+import Menu from '../Menu/Menu';
+import MenuItem from '../Menu/MenuItem';
+import MenuList from '../Menu/MenuList';
 import { mount } from 'enzyme';
 import { Popover } from '../Popover/Popover';
 import React from 'react';
 import renderer from 'react-test-renderer';
-import { Menu, MenuItem, MenuList } from '../Menu/Menu';
 
 describe('<Dropdown />', () => {
     const defaultMenu = (
