+++ src/Menu/MenuItem.js
@@ -2,51 +2,7 @@ import classnames from 'classnames';
 import PropTypes from 'prop-types';
 import React from 'react';
 
-// ------------------------------------------- Menu ------------------------------------------
-export const Menu = ({ addonBefore, children, className, ...props }) => {
-    const menuClasses = classnames(
-        'fd-menu',
-        {
-            'fd-menu--addon-before': addonBefore
-        },
-        className
-    );
-
-    return (
-        <nav {...props} className={menuClasses}>
-            {children}
-        </nav>
-    );
-};
-Menu.displayName = 'Menu';
-
-Menu.propTypes = {
-    addonBefore: PropTypes.bool,
-    className: PropTypes.string
-};
-
-Menu.propDescriptions = {
-    addonBefore: 'Set to **true** enables menu items with add-on before.'
-};
-
-// ---------------------------------------- Menu List ----------------------------------------
-export const MenuList = ({ children, className, ...props }) => {
-    const menuListClasses = classnames(
-        'fd-menu__list',
-        className
-    );
-
-    return <ul {...props} className={menuListClasses}>{children}</ul>;
-};
-MenuList.displayName = 'MenuList';
-
-MenuList.propTypes = {
-    children: PropTypes.node,
-    className: PropTypes.string
-};
-
-// ---------------------------------------- Menu Item ----------------------------------------
-export const MenuItem = ({ url, isLink, separator, addon, children, onclick, className, addonProps, urlProps, ...props }) => {
+const MenuItem = ({ url, isLink, separator, addon, children, onclick, className, addonProps, urlProps, ...props }) => {
     const menuItemLinkClasses = classnames(
         'fd-menu__item',
         {
@@ -93,6 +49,7 @@ export const MenuItem = ({ url, isLink, separator, addon, children, onclick, cla
         </React.Fragment>
     );
 };
+
 MenuItem.displayName = 'MenuItem';
 
 MenuItem.propTypes = {
@@ -115,24 +72,4 @@ MenuItem.propDescriptions = {
     urlProps: 'Additional props to be spread to the Menu Item links (when using `url`).'
 };
 
-// ---------------------------------------- Menu Group ----------------------------------------
-export const MenuGroup = ({ title, children, className, titleProps, ...props }) => {
-    const menuGroupClasses = classnames(
-        'fd-menu__group',
-        className
-    );
-
-    return (
-        <div {...props} className={menuGroupClasses}>
-            <h1 {...titleProps} className='fd-menu__title'>{title}</h1>
-            {children}
-        </div>
-    );
-};
-MenuGroup.displayName = 'MenuGroup';
-
-MenuGroup.propTypes = {
-    title: PropTypes.string.isRequired,
-    className: PropTypes.string,
-    titleProps: PropTypes.object
-};
+export default MenuItem;
