+++ src/index.js
@@ -31,7 +31,10 @@ export { default as InputGroup } from './InputGroup/InputGroup';
 export { default as FormGroup } from './InputGroup/FormGroup';
 export { ListGroup, ListGroupItem, ListGroupItemActions, ListGroupItemCheckbox } from './ListGroup/ListGroup';
 export { LocalizationEditor } from './LocalizationEditor/LocalizationEditor';
-export { Menu, MenuList, MenuItem, MenuGroup } from './Menu/Menu';
+export { default as Menu } from './Menu/Menu';
+export { default as MenuList } from './Menu/MenuList';
+export { default as MenuItem } from './Menu/MenuItem';
+export { default as MenuGroup } from './Menu/MenuGroup';
 export { Modal } from './Modal/Modal';
 export { MultiInput } from './MultiInput/MultiInput';
 export { Pagination } from './Pagination/Pagination';
